feat: add conflict handling for version deletion based on stock availability

+++ src/versions/versions.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { CreateVersionDto } from './dto/create-version.dto';
 import { UpdateVersionDto } from './dto/update-version.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -9,6 +9,7 @@ export class VersionsService {
 
   constructor( private readonly prismaORM : PrismaService ) {}
 
+  
 
   create(createVersionDto: CreateVersionDto) {
     
@@ -112,20 +113,63 @@ export class VersionsService {
 
   async remove(id: number) {
 
-    return this.prismaORM.$transaction( async (tx) => {
+    const errors : string[] = []
+
+    const transaction$ = this.prismaORM.$transaction( async (tx) => {
 
       const version = await tx.versions.findFirstOrThrow(
         {
           where: {id}
         }
       )
+      
+
+      const carsCount = await tx.cars.count(
+        {
+          where: {
+            model_name: version.model_name,
+            version_name: version.name
+          }
+        }
+      )
+
+
+      if( carsCount > 1 ) {
+
+        errors.push(`Version '${version.name}' has more than 1 unit in stock and can't be deleted directly` )
+
+        throw new ConflictException( errors )
+
+      }
+
+
+      const stock = await tx.cars.aggregate(
+        {
+          where: {
+            model_name: version.model_name,
+            version_name: version.name
+          },
+          _sum: {quantity: true}
+        }
+      )
+
+
+      const totalUnits = stock._sum.quantity ?? 0 ;
+
+      if( totalUnits > 1 ){
+
+        errors.push(  `Version '${version.name}' has more than 1 unit in stock and can't be deleted` )
+
+        throw new ConflictException( errors )
+
+      }
 
 
       await tx.cars.deleteMany(
         {
           where: {
             model_name: version.model_name,
-            version_name: version.name
+            version_name: version.name,
           }
         }
       )
@@ -136,6 +180,9 @@ export class VersionsService {
       })
 
     } )
+
+
+    return transaction$
 
   }
